--- src/app/components/ComunView/ComunView.component.html
@@ -1,84 +1,0 @@
-<app-header></app-header>
-<div class="container encabezado">
-    <div class="row mt-4">
-       <h1 class="text-white col-12">Comunicados Mercadology</h1>
-    </div>
-</div>
-
-<div class="container mt-2">
-
-  
-
-    <div style="background: #242424;" class="mat-elevation-z8 table-responsive">
-        <table style="background: #242424;" mat-table [dataSource]="dataSource" matSort class="bg-dark">
-
-            <!-- Position Column -->
-            <ng-container style="background: #242424;" matColumnDef="#" style="display:none;">
-                
-                <th mat-header-cell *matHeaderCellDef style="display:none; color: black !important;"> # </th>
-                <td class="contenedor-img-cs" mat-cell *matCellDef="let element; let i = index"> {{(i+1) +
-                    (paginator.pageIndex*paginator.pageSize)}}
-                    <div class="col-md-3 negro">
-                        <div class="contenedor-img-cs">
-                            <img src="../../../assets/imagenes/imagenPerfil/merca.jpeg" class=" img-cs " alt="imagen usuario">
-                        </div>
-                    </div>
-                
-                </td>
-            </ng-container>
-
-            <!-- Cliente Column -->
-            <ng-container matColumnDef="titulo">
-                <th mat-header-cell *matHeaderCellDef mat-sort-header class="tite so"><span cwlass="tite so" style="color: black !important;"> Titulo</span> </th>
-                <td mat-cell *matCellDef="let element" class="tite">
-                    {{element.titulo}}</td>
-            </ng-container>
-
-            <!-- Dominio Column -->
-            <ng-container matColumnDef="descripcion">
-                <th mat-header-cell *matHeaderCellDef mat-sort-header class="so"style="color: black !important;"> Descripcion </th>
-                <td mat-cell *matCellDef="let element" class="desc po"> {{element.descripcion}} </td>
-            </ng-container>
-
-            <!-- IP Column -->
-            <ng-container matColumnDef="fecha">
-                <th mat-header-cell *matHeaderCellDef mat-sort-header class="so" style="color: black !important;">Fecha</th>
-                <td mat-cell *matCellDef="let element" class="fecha po" style="color: black !important;"> <span>Publicado el:</span>    {{element.fecha}} </td>
-            </ng-container>
-
-            <!-- Estado Column -->
-            <ng-container matColumnDef="quien">
-                <th mat-header-cell *matHeaderCellDef mat-sort-header class="so" style="color: black !important;">Usuario/th>
-                <td mat-cell *matCellDef="let element" class="quien po">  {{element.quien}} </td>
-            </ng-container>
-
-            <ng-container style="display: none !important;" matColumnDef="ids">
-                <th mat-header-cell *matHeaderCellDef class="so" style="text-align: center; color: black !important;" > Opciones </th>
-                <!-- <td> {{element.ids}} </td> -->
-                <td mat-cell *matCellDef="let element; let i = index" style="display: none !important;" class="opciones" style="padding:0">
-          
-                    <!-- *ngIf="this.rol === '12345'"-->
-                    <button  *ngIf="this.rol === '12345'" class="btn btn-danger ms-3 btn-sm" style="display: none !important; width: 1% !important;"
-                        (click)="borrarComunicado( element, (i) + (paginator.pageIndex*paginator.pageSize) ) ">
-                        <i class="bi bi-trash-fill"></i>
-                    </button>
-                </td>
-            </ng-container>
-
-
-            <tr mat-header-row *matHeaderRowDef="displayedColumns" class="bg-light"></tr>
-            <tr mat-row *matRowDef="let row; columns: displayedColumns;"></tr>
-
-            <!-- Row shown when there is no matching data. -->
-            <tr class="mat-row" *matNoDataRow>
-                <td class="mat-cell" colspan="4">No hay datos de acuerdo a la búsqueda "{{input.value}}"</td>
-            </tr>
-
-        </table>
-
-
-        <!-- showFirstLastButtons -->
-    </div>
-    <mat-paginator style="background: #242424;" [pageSizeOptions]="[10, 25, 50, 100]" showFirstLastButtons></mat-paginator>
-</div>
-
