+++ src/app/components/comunicados/comunicados.component.css
@@ -1,59 +1,333 @@
-.title-cs {
-    color: #fff;
-    font-size: 30px;
-}
-
-.contenedor-cs {
-    display: flex;
-    justify-content: center;
-    flex-wrap: wrap;
-}
-
-.card-cs {
+@media (max-width: 480px){
+    .encabezado .row{
+        display: flex;
+        flex-direction: column;
+    }
+  }
+  .title-cs {
+  color: #fff;
+  font-size: 30px;
+  }
+  .contenedor-img-cs{
+  padding: 50px;
+  }
+  .text-white{
+  margin: center;
+  }
+  .contenedor-cs {
+  display: flex;
+  justify-content: center;
+  flex-wrap: wrap;
+  }
+  
+  .card-cs {
+  width: 80%;
+  border: 1px solid #242424;
+  font-family: "Poppins";
+  }
+  
+  .card-title {
+  font-family: "Helvetica-Bold";
+  }
+  
+  .card-text {
+  text-align: justify;
+  }
+  
+  .text-muted {
+  padding: 0;
+  margin: 0;
+  }
+  
+  .descargar-cs {
+  color: #FC6434;
+  text-decoration: none;
+  margin: 0;
+  padding: 0;
+  }
+  
+  .descargar-cs:hover {
+  color: #FCC42C;
+  text-decoration: none;
+  }
+  
+  .contenedor-img-cs {
+  display: flex;
+  justify-content: center;
+  flex-direction: column;
+  background: #242424;
+  padding-bottom: 30px;
+  padding-top: 10px;
+  }
+  thead{
+  background: #242424 !important;
+  }
+  tr.mat-header-row {
+  
+  height: 5px !important;
+  }
+  .img-cs {
+  border-radius: 200px;
+  width: 150px;
+  height: 150px;
+  }
+  
+  .eliminar {
+  background-color: #FC041C;
+  }
+  
+  .actualizar {
+  background-color: #646464;
+  }
+  
+  .btn-cs {
+  width: 8%;
+  float: right;
+  color: #fff;
+  margin-left: 10px;
+  margin-top: 10px;
+  }
+  table {
     width: 100%;
-    border: 1px solid #242424;
-    font-family: "Poppins";
-}
-
-.card-title {
-    font-family: "Helvetica-Bold";
-}
-
-.card-text {
-    text-align: justify;
-}
-.text-muted {
-    padding: 0;
-    margin: 0;
-}
-
-.descargar-cs {
-    color: #FC6434;
-    text-decoration: none;
-    margin: 0;
-    padding: 0;
-}
-
-.descargar-cs:hover {
-    color: #FCC42C;
-    text-decoration: none;
-}
-
-.contenedor-img-cs {
-    display: flex;
-    justify-content: center;
-    flex-direction: column;
-    background: #242424;
-    padding-bottom: 10px;
-    padding-top: 10px;
-}
-
-.img-cs {
-    border-radius: 200px;
-    width: 150px;
-    height: 150px;
-}
-
-#upt{
-    background: rgb(29, 160, 29) !important;
-}
+    color: white;
+  }
+  
+  .buscador {
+    font-size: 18px;
+  }
+  .buscador input{
+      color: white;
+      border: 0;
+      outline: none;
+      border-bottom: solid 2px white;
+      width: 100%;
+  }  
+  
+  th.mat-header-cell, td.mat-cell, td.mat-footer-cell{
+   
+    padding-right: 5px;
+    padding-left: 5px;
+  }
+  
+  @media (max-width: 767px){
+    .opciones{
+        display: flex;
+        justify-content: center;
+    }    
+  }
+  .opciones{        
+    text-align: center;
+  }
+  .holi{
+  padding: 150px;
+  }
+  @media (min-width: 1200px){
+  .container{
+    max-width: 80% !important;
+  }
+  }
+  .tite{
+  background: #FCC42C;
+  position: absolute;
+  margin-left: 22%;
+  margin-top: -220px;
+  border-radius: 5PX;
+  padding: 1PX;
+  box-sizing: border-box;
+  }
+  .so{
+  display: none;
+  }
+  .descripcion{
+  color: #242424 ;
+  }
+  .desc{
+    margin-top: 10px;
+    margin-left: 0px;
+    border-radius: 15PX;
+    background: whitesmoke  ;
+    border: 20px solid #242424;
+  }
+  .fecha{
+  border-radius: 15px;
+  position: absolute;
+  margin-right: -10%;
+  left: 65%;
+  margin-top: 200px !important;
+  background: #FCC42C;
+  color: wheat !important;
+  }
+  .quien{
+  position:absolute;
+    left: 16.9%;
+    margin-top: 200px;
+    box-sizing: border-box;
+    font-size: 20px;
+    color: wheat;
+  }@media (max-width: 480px){
+    .encabezado .row{
+        display: flex;
+        flex-direction: column;
+    }
+  }
+  .title-cs {
+  color: #fff;
+  font-size: 30px;
+  }
+  .contenedor-img-cs{
+  padding: 50px;
+  }
+  .text-white{
+  margin: center;
+  }
+  .contenedor-cs {
+  display: flex;
+  justify-content: center;
+  flex-wrap: wrap;
+  }
+  
+  .card-cs {
+  width: 80%;
+  border: 1px solid #242424;
+  font-family: "Poppins";
+  }
+  
+  .card-title {
+  font-family: "Helvetica-Bold";
+  }
+  
+  .card-text {
+  text-align: justify;
+  }
+  
+  .text-muted {
+  padding: 0;
+  margin: 0;
+  }
+  
+  .descargar-cs {
+  color: #FC6434;
+  text-decoration: none;
+  margin: 0;
+  padding: 0;
+  }
+  
+  .descargar-cs:hover {
+  color: #FCC42C;
+  text-decoration: none;
+  }
+  
+  .contenedor-img-cs {
+  display: flex;
+  justify-content: center;
+  flex-direction: column;
+  background: #242424;
+  padding-bottom: 30px;
+  padding-top: 10px;
+  }
+  thead{
+  background: #242424 !important;
+  }
+  tr.mat-header-row {
+  
+  height: 5px !important;
+  }
+  .img-cs {
+  border-radius: 200px;
+  width: 150px;
+  height: 150px;
+  }
+  
+  .eliminar {
+  background-color: #FC041C;
+  }
+  
+  .actualizar {
+  background-color: #646464;
+  }
+  
+  .btn-cs {
+  width: 8%;
+  float: right;
+  color: #fff;
+  margin-left: 10px;
+  margin-top: 10px;
+  }
+  table {
+    width: 100%;
+    color: white;
+  }
+  
+  .buscador {
+    font-size: 18px;
+  }
+  .buscador input{
+      color: white;
+      border: 0;
+      outline: none;
+      border-bottom: solid 2px white;
+      width: 100%;
+  }  
+  
+  th.mat-header-cell, td.mat-cell, td.mat-footer-cell{
+   
+    padding-right: 5px;
+    padding-left: 5px;
+  }
+  
+  @media (max-width: 767px){
+    .opciones{
+        display: flex;
+        justify-content: center;
+    }    
+  }
+  .opciones{        
+    text-align: center;
+  }
+  .holi{
+  padding: 150px;
+  }
+  @media (min-width: 1200px){
+  .container{
+    max-width: 80% !important;
+  }
+  }
+  .tite{
+  background: #FCC42C;
+  position: absolute;
+  margin-left: 22%;
+  margin-top: -220px;
+  border-radius: 5PX;
+  padding: 1PX;
+  box-sizing: border-box;
+  }
+  .so{
+  display: none;
+  }
+  .descripcion{
+  color: #242424 ;
+  }
+  .desc{
+    margin-top: 10px;
+    margin-left: 0px;
+    border-radius: 15PX;
+    background: whitesmoke  ;
+    border: 20px solid #242424;
+  }
+  .fecha{
+  border-radius: 15px;
+  position: absolute;
+  margin-right: -10%;
+  left: 65%;
+  margin-top: 200px !important;
+  background: #FCC42C;
+  color: wheat !important;
+  }
+  .quien{
+  position:absolute;
+    left: 16.9%;
+    margin-top: 200px;
+    box-sizing: border-box;
+    font-size: 20px;
+    color: wheat;
+  }
