+++ src/app/components/header/header.component.html
@@ -11,10 +11,10 @@
         <div class="collapse navbar-collapse justify-content-end" id="navbarNav">
             <ul class="navbar-nav">
                 <li class="nav-item">
-                    <a [routerLink]="['/ComunView']" routerLinkActive="router-link-active" class="nav-link" aria-current="page">Comunicados</a>
+                    <a [routerLink]="['/comunicados']" routerLinkActive="router-link-active" class="nav-link" aria-current="page">Comunicados</a>
                 </li>
                 <li class="nav-item" *ngIf="this.rol === '12345' || this.rol === '234567' || this.puesto == 'Manager' || this.puesto == 'Director'">
-                    <a [routerLink]="['/MisComun']" routerLinkActive="router-link-active" class="nav-link" aria-current="page">MisComunicados</a>
+                    <a [routerLink]="['/MisComun']" routerLinkActive="router-link-active" class="nav-link" aria-current="page">Mis Comunicados</a>
                 </li>
                 <li class="nav-item" *ngIf="this.rol === '12345' || this.rol === '234567'">
                     <a [routerLink]="['/usuarios']" routerLinkActive="router-link-active" class="nav-link" aria-current="page">Gestión de usuarios</a>
