+++ src/app/components/miscomunicados/miscomunicados.component.css
@@ -1,39 +1,39 @@
 @media (max-width: 480px){
-    .encabezado .row{
-        display: flex;
-        flex-direction: column;
-    }
+  .encabezado .row{
+      display: flex;
+      flex-direction: column;
+  }
 }
 
 table {
-    width: 100%;
+  width: 100%;
+  color: white;
+}
+
+.buscador {
+  font-size: 18px;
+}
+.buscador input{
     color: white;
-  }
-  
-  .buscador {
-    font-size: 18px;
-  }
-  .buscador input{
-      color: white;
-      border: 0;
-      outline: none;
-      border-bottom: solid 2px white;
-      width: 100%;
-  }  
+    border: 0;
+    outline: none;
+    border-bottom: solid 2px white;
+    width: 100%;
+}  
 
-  th.mat-header-cell, td.mat-cell, td.mat-footer-cell{
-    color: white !important;
-    padding-right: 5px;
-    padding-left: 5px;
-  }
+th.mat-header-cell, td.mat-cell, td.mat-footer-cell{
+  color: white !important;
+  padding-right: 5px;
+  padding-left: 5px;
+}
 
-  @media (max-width: 767px){
-    .opciones{
-        display: flex;
-        justify-content: center;
-    }    
+@media (max-width: 767px){
+  .opciones{
+      display: flex;
+      justify-content: center;
+  }    
 }
 .opciones{        
-    width: 150px!important;
-    text-align: center;
+  width: 150px!important;
+  text-align: center;
 }
