+++ src/app/components/miscomunicados/miscomunicados.component.html
@@ -9,7 +9,7 @@
 
     <div class="row" style="justify-content: space-between;">
         <div class="col-lg-2 mb-4">
-            <button routerLink="/comunicados/add" class="btn btn-primary col-lg-12">
+            <button routerLink="/comunicado/add" class="btn btn-primary col-lg-12">
                 <i class="bi bi-cloud-arrow-up-fill me-1"></i> Agregar
             </button>
         </div>
@@ -57,11 +57,11 @@
                 <th mat-header-cell *matHeaderCellDef style="text-align: center; color: black !important;" > Opciones </th>
                 <!-- <td> {{element.ids}} </td> -->
                 <td mat-cell *matCellDef="let element; let i = index" class="opciones" style="padding:0">
-                    <button class="btn btn-primary btn-sm" [routerLink]="['/comunicados', element.ids ]">
+                    <button class="btn btn-primary btn-sm" [routerLink]="['/comunicado', element.ids ]">
                         <i class="bi bi-info-circle-fill"></i>
                     </button>
 
-                    <button class="btn btn-secondary ms-3 btn-sm" [routerLink]="['/comunicados', element.ids ]">
+                    <button class="btn btn-secondary ms-3 btn-sm" [routerLink]="['/comunicado', element.ids ]">
                         <i class="bi bi-pencil-square"></i>
                     </button>
