+++ src/app/services/auth.service.ts
@@ -165,7 +165,7 @@ export class AuthService {
         this.regreso = "Administrador";
       } else if (sessionStorage.getItem('rol') == "234567") {
         this.regreso = "Editor";
-      }else if(sessionStorage.getItem('rol') == undefined || sessionStorage.getItem('rol') == ""){
+      } else if (sessionStorage.getItem('rol') == undefined || sessionStorage.getItem('rol') == "") {
         this.regreso = undefined;
         this.Logout();
       }
@@ -359,7 +359,7 @@ export class AuthService {
         })
       );
   }
-  
+
   UpdatCliente(cliente: ClienteModel) {
     const token = sessionStorage.getItem('token');
     const ClienteTemp = {
@@ -391,13 +391,23 @@ export class AuthService {
   getComunicado(ids: string) {
     return this.http.get(`${this.url}/comunicados/${ids}.json`);
   }
-  getComun() {
-    return this.http.get(`${this.url}/comunicados.json`)
-    .pipe(
-      map(this.CrearComun),
-      delay(1500)
-    );
-     
+  //Tipo se refiere a si se va a kostrar en comunicados general (donde se visualizan) o si es donde se eliminan, agregan o editan comunicados
+  getComun(tipo: string) {
+
+    if (tipo == "privado") {
+      return this.http.get(`${this.url}/comunicados.json`)
+      .pipe(
+        map(this.CrearComunPrivado),
+        delay(1500)
+      );
+    } else {
+      return this.http.get(`${this.url}/comunicados.json`)
+        .pipe(
+          map(this.CrearComun),
+          delay(1500)
+        );
+    }
+
   }
   DeleteComun(ids: string) {
     const token = sessionStorage.getItem('token');
@@ -419,13 +429,28 @@ export class AuthService {
     let date = new Date();
     let mes = date.getMonth() + 1;
     let meses;
-    if(mes < 10){
-       meses = `0${date.getMonth() + 1}`;
+    if (mes < 10) {
+      meses = `0${date.getMonth() + 1}`;
     }
     let hoy = `${date.getFullYear()}-${meses}-${date.getDate()}`;
-    return Comunicado.filter((({fecha}) => fecha <= hoy));
+    return Comunicado.filter((({ fecha }) => fecha <= hoy));
 
     // return Comunicado;
+  }
+
+  private CrearComunPrivado(ComunicadoObj: object) {
+    const Comunicado: ComunicadoModel[] = [];
+    if (ComunicadoObj === null) {
+      return [];
+    }
+
+    Object.keys(ComunicadoObj).forEach(key => {
+      const comunicado: ComunicadoModel = ComunicadoObj[key];
+      comunicado.ids = key;
+      Comunicado.push(comunicado);
+    });
+
+    return Comunicado;
   }
 
   getShow(ids: string) {
@@ -444,7 +469,7 @@ export class AuthService {
     const Cliente: ClienteModel[] = [];
     // console.log(ClienteObj);
     if (ClienteObj === null) {
-      return []; 
+      return [];
     }
 
     Object.keys(ClienteObj).forEach(key => {
@@ -624,7 +649,7 @@ export class AuthService {
 
   /**Agrega a accesos**/
   agregarAccesos(acceso: AccesosModel, token: string) {
-    
+
     let authData = {
       ...acceso
     };
@@ -640,8 +665,8 @@ export class AuthService {
   };/**Cierra el Registrar datos accesos**/
 
 
-   /**Hacer un get de todos los usuarios**/
-   getAccesos() {
+  /**Hacer un get de todos los usuarios**/
+  getAccesos() {
     return this.http.get(`${this.urlDatos}/acceso.json`)
       .pipe(
         map(this.crearArregloAcceso)
@@ -650,7 +675,7 @@ export class AuthService {
 
   private crearArregloAcceso(accesoObj: object) {
     const accesos: AccesosModel[] = [];
-    
+
     if (accesoObj === null) { return []; }
 
     Object.keys(accesoObj).forEach(key => {
@@ -662,7 +687,7 @@ export class AuthService {
   }
 
   modificarAcceso(acceso: AccesosModel, token: string) {
-    
+
     const AccesoTemp = {
       ...acceso
     };
@@ -675,12 +700,12 @@ export class AuthService {
     return this.http.delete(`${this.urlDatos}/acceso/${id}.json` + this.auth + token);
   }
 
-  
+
 
 
   /** Agregar comunicados **/
   // agregarComunicado(token:string, comunicado:ComunicadoModel){
-    
+
   //   let comunicadoDatos = {
   //     ...comunicado
   //   };
@@ -701,7 +726,7 @@ export class AuthService {
 
   // private crearArregloComunicado(comunicadoObj: object) {
   //   const comunicados: ComunicadoModel[] = [];
-    
+
   //   if (comunicadoObj === null) { return []; }
 
   //   Object.keys(comunicadoObj).forEach(key => {
@@ -713,7 +738,7 @@ export class AuthService {
   // }
 
   // modificarComunicado(comunicado: ComunicadoModel, token: string) {
-    
+
   //   const ComunicadoTemp = {
   //     ...comunicado
   //   };
